+++ src/memlord/schemas/tools/__init__.py
@@ -1,5 +1,5 @@
 from .delete import DeleteResult
-from .list_memories import MemoryDetail, MemoryItem, MemoryListItem, MemoryPage
+from .list_memories import MemoryDetail, MemoryItem, MemoryPage
 from .recall import RecallPage, RecallResult
 from .retrieve import MemoryResult
 from .store import StoreResult
