+++ src/memlord/schemas/tools/list_memories.py
@@ -6,23 +6,6 @@ from ..memory_type import MemoryType
 from ..pagination import Paginated
 
 
-class MemoryListItem(BaseModel):
-    """Full memory record — used by UI and DAO."""
-
-    id: int
-    name: str
-    content: str
-    memory_type: MemoryType
-    metadata: dict = Field(default_factory=dict)
-    tags: set[str]
-    created_at: NaiveDatetime
-    workspace_id: int
-
-    @field_serializer("created_at")
-    def serialize_created_at(self, v: datetime) -> str:
-        return v.replace(tzinfo=UTC).isoformat()
-
-
 class MemoryItem(BaseModel):
     """Slim memory record returned by MCP list/search tools (no id, no content)."""
