+++ src/memlord/tools/search_by_tag.py
@@ -10,7 +10,7 @@ from memlord.dao import MemoryDao
 from memlord.dao.workspace import WorkspaceDao
 from memlord.db import MCPSessionDep
 from memlord.models import Memory, MemoryTag, Tag, Workspace
-from memlord.schemas import MemoryItem, MemoryPage
+from memlord.schemas.tools import MemoryItem, MemoryPage
 
 mcp = FastMCP()
